move traced implementation to library package
